Remove circular dependency
