+++ packages/mds-db/package.json
@@ -22,7 +22,8 @@
     "@mds-core/mds-logger": "0.1.24",
     "@mds-core/mds-types": "0.1.23",
     "@mds-core/mds-utils": "0.1.26",
-    "pg": "8.0.2"
+    "pg": "8.0.2",
+    "uuid": "7.0.3"
   },
   "devDependencies": {
     "@mds-core/mds-test-data": "0.1.26"
