+++ packages/mds-db/tests/mds-db.spec.ts
@@ -10,7 +10,6 @@ import {
   makeDevices,
   makeEventsWithTelemetry,
   makeEvents,
-  makeTrip,
   JUMP_PROVIDER_ID,
   POLICY_JSON,
   POLICY2_JSON,
@@ -20,9 +19,10 @@ import {
   LA_CITY_BOUNDARY,
   DISTRICT_SEVEN
 } from '@mds-core/mds-test-data'
-import { now, clone, NotFoundError } from '@mds-core/mds-utils'
+import { now, clone, NotFoundError, rangeRandomInt } from '@mds-core/mds-utils'
 
 import { isNullOrUndefined } from 'util'
+import { v4 as uuid } from 'uuid'
 import MDSDBPostgres from '../index'
 
 import { dropTables, createTables, updateSchema } from '../migration'
@@ -50,6 +50,42 @@ const DistrictSeven: Geography = {
   name: 'District Seven',
   geography_id: GEOGRAPHY2_UUID,
   geography_json: DISTRICT_SEVEN
+}
+
+function makeTrip(device: Device): Trip {
+  return {
+    provider_id: device.provider_id,
+    provider_name: device.provider_id,
+    device_id: device.device_id,
+    vehicle_id: device.vehicle_id,
+    vehicle_type: device.type,
+    propulsion_type: device.propulsion,
+    provider_trip_id: uuid(),
+    trip_duration: rangeRandomInt(5),
+    trip_distance: rangeRandomInt(5),
+    route: {
+      type: 'FeatureCollection',
+      features: [
+        {
+          type: 'Feature',
+          properties: {
+            timestamp: now()
+          },
+          geometry: {
+            type: 'Point',
+            coordinates: [Math.random() * 10, Math.random() * 10]
+          }
+        }
+      ]
+    },
+    accuracy: Math.random() * 3,
+    trip_start: now() - 1000 * Math.random(),
+    trip_end: now(),
+    parking_verification_url: 'http://iamverified.com',
+    standard_cost: rangeRandomInt(5),
+    actual_cost: rangeRandomInt(5),
+    recorded: now()
+  }
 }
 
 /* You'll need postgres running and the env variable PG_NAME
